requirements.txt
pypdf==6.8.0
pdfplumber==0.11.9
openpyxl==3.1.5
streamlit==1.55.0
pandas==2.3.3
gspread==6.2.1
google-auth==2.49.0
requests==2.32.5
